+++ nautobot_design_builder/tests/__init__.py
@@ -35,7 +35,7 @@ class DesignTestCase(TestCase):
     def get_mocked_job(self, design_class: Type[DesignJob]):
         """Create an instance of design_class and properly mock request and job_result for testing."""
         job = design_class()
-        job._setup_journal = lambda *args: None  # pylint: disable=protected-access
+        job._setup_journal = mock.Mock()
 
         job.job_result = mock.Mock()
         if nautobot_version < "2.0.0":
