+++ nautobot_design_builder/tests/test_model_design_instance.py
@@ -1,11 +1,14 @@
 """Test DesignInstance."""
 
+from unittest import mock
+import uuid
 from django.core.exceptions import ValidationError
 from django.db import IntegrityError
 from django.contrib.contenttypes.models import ContentType
 
-from nautobot.extras.models import Status
+from nautobot.extras.models import Status, JobResult, Job
 
+from nautobot_design_builder.util import nautobot_version
 
 from .test_model_design import BaseDesignTest
 from .. import models, choices
@@ -28,6 +31,24 @@ class BaseDesignInstanceTest(BaseDesignTest):
         )
         design_instance.validated_save()
         return design_instance
+
+    def create_journal(self, job, design_instance, kwargs):
+        """Creates a Journal."""
+        job_result = JobResult(
+            job_model=self.job1,
+            name=job.class_path,
+            job_id=uuid.uuid4(),
+            obj_type=ContentType.objects.get_for_model(Job),
+        )
+        job_result.log = mock.Mock()
+        if nautobot_version < "2.0":
+            job_result.job_kwargs = {"data": kwargs}
+        else:
+            job_result.task_kwargs = kwargs
+        job_result.validated_save()
+        journal = models.Journal(design_instance=design_instance, job_result=job_result)
+        journal.validated_save()
+        return journal
 
     def setUp(self):
         super().setUp()
